# api/api-docs.yaml
openapi: 3.0.0
info:
  title: Greenalytic User Management API
  version: 1.0.0
  description: API for managing users with authentication and pagination
servers:
  - url: http://localhost:5000/greenalytic
    description: Local Development Server
  - url: https://greenalytic-vehicle-monitoring-1.onrender.com/greenalytic
    description: Production Server

# Define Authentication Scheme
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  # JSON Web Token (JWT) format
  schemas:
...
    EmissionData:
      type: object
      properties:
        id:
          type: integer
          example: 1
        co2Percentage:
          type: number
          example: 4.5
        coPercentage:
          type: number
          example: 0.3
        o2Percentage:
          type: number
          example: 21.0
        hcPPM:
          type: integer
          example: 200
        vehicleId:
          type: integer
          example: 101
        plateNumber:
          type: string
          example: RAB123X
        trackingDeviceId:
          type: integer
          example: 5001
        timestamp:
          type: string
          format: date-time
          example: "2025-03-01T10:30:00Z"
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        totalCount:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10
paths:
  /users:
    get:
      summary: Get all users (Protected)
      description: Retrieve a paginated list of users. Requires authentication.
      security:
        - BearerAuth: []  # Require Bearer Token for this route
      parameters:
        - name: page
          in: query
          description: Page number (default is 1)
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of users per page (default is 10)
          required: false
          schema:
            type: integer
            default: 10
      tags: [authentication]
      responses:
        '200':
          description: A list of users with pagination details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        username:
                          type: string
                        email:
                          type: string
                        image:
                          type: string
                        role:
                          type: string
                        phoneNumber:
                          type: string
                        verified:
                          type: boolean
                        createdAt:
                          type: string
                          format: date-time
                        updatedAt:
                          type: string
                          format: date-time
                        vehicles:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: string
                              plateNumber:
                                type: string
                              vehicleModel:
                                type: string
                              vehicleType:
                                type: string
                        trackingDevices:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: string
                              serialNumber:
                                type: string
                              model:
                                type: string
                              type:
                                type: string
                              isActive:
                                type: boolean
                  pagination:
                    type: object
                    properties:
                      currentPage:
                        type: integer
                      totalPages:
                        type: integer
                      remainingItems:
                        type: integer
                      totalItems:
                        type: integer
                      limit:
                        type: integer
        '401':
          description: Unauthorized - Invalid or missing authentication token
        '500':
          description: Internal server error
  /users/signup:    
    post:
      summary: User Signup
      tags: [authentication]
      operationId: signup
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - username
                - phoneNumber
              properties:
                username:
                  type: string
                  example: "JohnDoe"
                email:
                  type: string
                  example: "imanariyobaptiste@gmail.com"
                password:
                  type: string
                  example: "123"
                phoneNumber:
                  type: string
                  example: "1234567890"
                location:
                  type: string
                  example: "City, Country"
                gender:
                  type: string
                  example: "Male"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: "User registered successfully"
                  userInformation:
                    type: object
                    properties:
                      id:
                        type: integer
                      email:
                        type: string
                      username:
                        type: string
                      phoneNumber:
                        type: string
                      role:
                        type: string
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: "Please provide all required fields: email, password, username, and phoneNumber."
        '500':
          description: Internal Server Error
  /users/login:     
    post:
      summary: User Login
      tags: [authentication]
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: "imanariyobaptiste@gmail.com"
                password:
                  type: string
                  example: "123"
      responses:
        '200':
          description: User logged in successfully, returns a JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: "Login successful"
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZXhwIjoxNjE0MzMwMjk5fQ.WD_Vn0Kg5XsOpbZTYrs2g2ev_JfBltJ_4y1M8RQ2c4w"
        '400':
          description: Invalid email or password
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: "Invalid email or password."
        '500':
          description: Internal Server Error
...
  /emissions:
    post:
      summary: Create a new emission data record
      security:
        - bearerAuth: []
      tags:
        - Emission Data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmissionData"
      responses:
        "201":
          description: Emission data created successfully
        "400":
          description: Bad Request - Missing required fields
        "500":
          description: Internal Server Error

    get:
      summary: Get all emission data (paginated)
      security:
        - bearerAuth: []
      tags:
        - Emission Data
      parameters:
        - name: page
          in: query
          schema:
            type: integer
          example: 1
        - name: limit
          in: query
          schema:
            type: integer
          example: 10
        - name: startTime
          in: query
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: List of emission data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/EmissionData"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
        "500":
          description: Internal Server Error
...